router/pyproject.toml
[project]
name = "router"
version = "0.1.0"
description = "Router Agent Project"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.2.5",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.4",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "arize-phoenix-otel==0.10.1",
    "attrs==25.3.0",
    "Authlib==1.6.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cryptography==45.0.3",
    "Deprecated==1.2.18",
    "distro==1.9.0",
    "docstring_parser==0.16",
    "dotenv==0.9.9",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "google-adk==1.1.1",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.95.1",
    "google-cloud-bigquery==3.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.23.3",
    "google-cloud-speech==2.32.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.1",
    "google-crc32c==1.7.1",
    "google-genai==1.17.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.73.0rc1",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.2",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.32.3",
    "idna==3.10",
    "importlib_metadata==8.6.1",
    "Jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "litellm",
    "MarkupSafe==3.0.2",
    "mcp==1.9.2",
    "multidict==6.4.4",
    "numpy==2.2.6",
    "openai==1.82.1",
    "openinference-instrumentation==0.1.32",
    "openinference-semantic-conventions==0.1.17",
    "opentelemetry-api==1.33.1",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-exporter-otlp==1.33.1",
    "opentelemetry-exporter-otlp-proto-common==1.33.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.33.1",
    "opentelemetry-exporter-otlp-proto-http==1.33.1",
    "opentelemetry-proto==1.33.1",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "packaging==25.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "pydantic_core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.41",
    "sse-starlette==2.3.5",
    "starlette==0.46.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.13.2",
    "tzlocal==5.3.1",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "pytest-asyncio>=0.23.7",
    "respx>=0.21.0",
    "pytest-cov>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20"
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
litellm = { path = "litellm-1.72.3-py3-none-any.whl" }

[build-system]
requires = ["hatchling", "setuptools"]
build-backend = "hatchling.build"
 